+++ scripts/generate_minknow_samplesheet.py
@@ -17,7 +17,7 @@ from tabulate import tabulate
 
 from data.ONT_barcodes import ONT_BARCODE_LABEL_PATTERN, ONT_BARCODES
 from scilifelab_epps.epp import traceback_to_step, upload_file
-from scilifelab_epps.epp_utils.udf_tools import fetch
+from scilifelab_epps.utils.udf_tools import fetch
 from scilifelab_epps.wrapper import epp_decorator
 
 DESC = """ Script to generate MinKNOW samplesheet for starting ONT runs.
