+++ scripts/molar_concentration.py
@@ -16,7 +16,7 @@ from genologics.config import BASEURI, PASSWORD, USERNAME
 from genologics.entities import Process
 from genologics.lims import Lims
 
-from scilifelab_epps.epp_utils.formula import ng_ul_to_nM
+from scilifelab_epps.utils.formula import ng_ul_to_nM
 from scilifelab_epps.epp import EppLogger
